% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPRISMNormal}
\alias{getPRISMNormal}
\title{PRISM Normal Download Function}
\usage{
getPRISMNormal(type, month = NA, spGrid = TRUE)
}
\arguments{
\item{type}{A string denoting the datatype to be downloaded.
"tmax", "tmin", "ppt".}

\item{month}{Month of year to download data for. \code{NA}
indicates all months.}

\item{\code{logical}}{for \code{SpatialGridDataFrame} return value
vs storing in working directory. Default is \code{TRUE}.}
}
\value{
Either a SpatialGridDataFrame from the sp package
(spGrid==TRUE) or
the bil and associated files are stored in the working
directory and the function returns TRUE/FALSE depending on
the success of the unzipping.  The current naming
convention for the bil files (as of 12/13/2013) are for
daily: 'PRISM_type_stable_4kmD1_yyyymmdd_bil.bil' e.g.
'PRISM_tmax_stable_4kmD1_19990101_bil.bil' monthly: #
'PRISM_type_stable_4kmM2_yyyymm_bil.bil'
}
\description{
This function downloads a PRISM weather data
normal grid to disk or the R-environment. If downloading
to disk, .bil files are downloaded; if downloading
to the R-environment, a \code{SpatialGridDataFrame}
is downloaded. A \code{.RData} file containing download
info is also provided.
}
\details{
The data these scripts access are subject to the
PRISM Gridded Climate Data -
Terms of Use. As of 07/06/2014, these are available at
http://www.prism.oregonstate.edu/documents/PRISM_terms_of_use.pdf
NOTE in particular the Data section: 'All data (gridded,
polygon, tabular, graphical) retrieved from the website or
otherwise provided on the website may be freely reproduced
and distributed for non-commercial purposes.

When referring to the data, the source must be clearly and
prominently stated and include, at a minimum, our name,
URL, and the date of data creation. For example: PRISM
Climate Group, Oregon State University,
http://prism.oregonstate.edu, created 4 Feb 2004.

Sale or other commercial use of the data must be arranged by
contacting the PRISM Climate Group.'
}

